Restored Sublime default brace matching and tab behavior.  Enabled change tracking to handle case of snippets with text fields.  Enabled tab to move among snippet text fields without breaking change tracking. Restored default Sublime behavior for \n between paired curly braces (open line between the braces and indent).  Prepared to enable Sublime auto indent to work with change tracking.
Added ctrl+; and ctrl+}.  The former formats the current line and will
auto-indent on an empty line, placing the cursor at the correct
column (Emacs tab behavior).  The latter formats the innermost
block (pair of curly braces) and places the cursor outside the block so
that a second press of ctrl+} will format the next innermost block and
so on.