Prevent unnecessary calls to service.configure
The Sublime API is somewhat unclear, but the 'callback' argument for
Settings.add_on_change will be called whenever *any* setting changes.
The 'key' argument is simply a way of referring to the callback in a
subsequent call to Settings.clear_on_change, *not* the setting name
that will trigger the callback.

Two changes are included here:
- Only register on_change handler once, since it will be triggered on
every setting change
- Check previous format options against current values to detect a
change; only call 'configure' if one of the values has actually changed

Fixes #187